pyproject.toml
[project]
name = "backend"
version = "0.3.8"
description = "API WEB HOSPITAL-SDLG"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.10.1",
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.0.0",
    "bitsandbytes>=0.48.1",
    "cachetools>=5.5.2",
    "click>=8.1.8",
    "cryptography>=44.0.2",
    "faker>=37.3.0",
    "fastapi[all]>=0.115.12",
    "fastmcp>=2.12.4",
    "matplotlib>=3.10.6",
    "maturin>=1.9.6",
    "numpy>=2.3.3",
    "opencv-python>=4.11.0.86",
    "orjson>=3.10.16",
    "passlib>=1.7.4",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "polars>=1.32.2",
    "polars-lts-cpu>=1.33.1",
    "psycopg>=3.2.6",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.2",
    "pyjwt>=2.10.1",
    "pyrefly>=0.27.0",
    "pytesseract>=0.3.13",
    "pytest>=8.3.5",
    "pytest-rich>=0.2.0",
    "python-dotenv>=1.1.0",
    "python-magic>=0.4.27",
    "python-magic-bin>=0.4.14",
    "python-multipart>=0.0.20",
    "redis>=6.4.0",
    "requests>=2.32.4",
    "scalar-fastapi>=1.0.3",
    "sqlmodel>=0.0.24",
    "stripe>=12.4.0",
    "transformers[torch]>=4.56.2",
]

[tool]

[tool.pyrefly]
project-includes = ["**/*"]
project-excludes = ['**/*venv/**\*']
